core: Move FontDescriptor to font/font_descriptor.rs
This makes font.rs smaller, as it's a pretty big file.
